requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Pillow==10.1.0
numpy==1.26.2
